updated Assessment components hierarchy
