--- src/view/pages/logged-in/components/sm-employee/Assessment/Assessment.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { Categories } from "./Categories";
-import { CategorySkills } from './CategorySkills';
-
-function Assessment() {
-    return (
-        <React.Fragment>
-            <Categories className='sm-table sm-component'></Categories>
-            <CategorySkills className='sm-table sm-component'></CategorySkills>
-        </React.Fragment>
-    );
-}
-
-export { Assessment };
